testing/web-platform/tests/old-tests/submission/Infraware/Session_History/contents/Session_History/combination_history_002.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Session History</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>combination_history_002(After calling of pushState method, check length.)</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>
		
		try {
			var first;
			var second;
			first = window.history.length;
			window.history.pushState(1, document.title, '?x=1');
			second = window.history.length;

			test(function() {
				assert_equals(second - first, 1, "combination test is failed.");
			});
		} catch (e) {
			test(function() {
				assert_unreached("call of pushState method is failed.");
			});
		}

	</script>
 </body>
</html>
